japanese/requirements.txt
flask
openai-whisper
torch
jiwer
